[FIX] l10n_br_account: Account Mapping
Fix error when mapping accounts based on fiscal position. Now the
account mapping are undoing before the payment lines recompute and
doing after this. It is necessary because the recompute always do
the lines balance and as the account usually do not be a payable or
receivable account, some unnecessary line created.