[IMP] l10n_br_account
This commit intends to clarify the way that taxes retentions are made,
without break the way that already works. A new boolean field called
l10n_br_retention was added and based on it the tax amount is multiply
by -1. When the tax amount is negative, this field is set to true.